Fix for not being able to write to the working directory
When check_website runs wget it expects to be able to write to its working directory.  When it cannot wget returns an error and the check fails.  This patch writes the files to /dev/null so that they're never created in the first place.  This also means we don't have to pass the delete files flag.  Multiple people are having an issue with this as mentioned on the bottom of this page: http://exchange.nagios.org/directory/Plugins/Websites,-Forms-and-Transactions/nagios-2Dcheck-2Dwebsite/details